chore(yir): make the native review the only path
The native YIR/MIR pages are now always rendered, so drop the YearInReview and
MonthInReview feature flags and the iframed fallback entirely: the review routes
render YirPage/MirPage directly, the profile links that gated on the flag show
unconditionally, and the og iframe frame builders (yearToDate, monthInReview)
plus their slurm factory are removed from UrlBuilder.
