+++ projects/client/i18n/messages/bg-BG.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "От началото на годината",
   "preview_feature_description_edit_mode": "Активирайте контроли за скриване или показване на конфигурируеми секции в цялото приложение.",
   "preview_feature_description_leaderboard": "Класирайте хората, които следите, по време на гледане и вижте къде се намирате сред тях.",
-  "preview_feature_description_month_in_review": "Прегледайте Месечен преглед, който вече работи директно във V3.",
   "preview_feature_description_rewatch": "Започнете сесии за повторно гледане на предавания, които вече сте гледали, отбележете гледани епизоди директно от прозореца на сесията.",
   "preview_feature_description_scoped_favorites": "Ограничете във времето списъците с любими, като показвате любими от текущата година в профилите и групирате пълния списък с любими по година.",
   "preview_feature_description_smart_recommendations": "Използвайте нашия двигател от следващо поколение, за да откривате препоръчани филми и предавания отвъд очевидното.",
   "preview_feature_description_smart_related": "Използвайте нашия двигател от следващо поколение, за да откривате свързани филми и предавания отвъд очевидното.",
-  "preview_feature_description_year_in_review": "Прегледайте Годишен преглед, който вече работи директно във V3.",
   "preview_feature_title_edit_mode": "Режим на редактиране",
   "preview_feature_title_leaderboard": "Пантеон",
-  "preview_feature_title_month_in_review": "Месечен преглед",
   "preview_feature_title_rewatch": "Преглед",
   "preview_feature_title_scoped_favorites": "Обхванати любими",
   "preview_feature_title_smart_recommendations": "Интелигентни препоръки",
   "preview_feature_title_smart_related": "Интелигентно свързани",
-  "preview_feature_title_year_in_review": "Годишен преглед",
   "reset_status_cleared": "Списъкът за гледане е изчистен.",
   "section_title_season_progress": "Сезон {number} напредък",
   "section_title_seasons_overview": "Преглед",
