+++ projects/client/i18n/messages/ca-ES.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Acumulat de l'any",
   "preview_feature_description_edit_mode": "Habilita els controls per amagar o mostrar seccions configurables a tota l'aplicació.",
   "preview_feature_description_leaderboard": "Classifica les persones que segueixes per temps de visionat i mira on quedes entre elles.",
-  "preview_feature_description_month_in_review": "Previsualitza el Resum del Mes, ara executant-se directament en V3.",
   "preview_feature_description_rewatch": "Inicia sessions de revisió per a programes que ja has vist, fes un seguiment o marca els episodis com a vistos directament des del calaix de la sessió.",
   "preview_feature_description_scoped_favorites": "Delimita les llistes de preferits mostrant els preferits de l'any actual als perfils i agrupant la llista completa de preferits per any.",
   "preview_feature_description_smart_recommendations": "Utilitza el nostre motor de nova generació per trobar pel·lícules i sèries recomanades més enllà del que és obvi.",
   "preview_feature_description_smart_related": "Utilitza el nostre motor de nova generació per trobar pel·lícules i sèries relacionades més enllà del que és obvi.",
-  "preview_feature_description_year_in_review": "Previsualitza el Resum de l'Any, ara executant-se directament en V3.",
   "preview_feature_title_edit_mode": "Mode d'edició",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Resum del Mes",
   "preview_feature_title_rewatch": "Revisió",
   "preview_feature_title_scoped_favorites": "Preferits per àmbit",
   "preview_feature_title_smart_recommendations": "Recomanacions intel·ligents",
   "preview_feature_title_smart_related": "Relacionats intel·ligents",
-  "preview_feature_title_year_in_review": "Resum de l'Any",
   "reset_status_cleared": "Llista de seguiment esborrada.",
   "section_title_season_progress": "Temporada {number} progrés",
   "section_title_seasons_overview": "Visió general",
