+++ projects/client/i18n/messages/da-DK.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "År til dato",
   "preview_feature_description_edit_mode": "Aktiver kontroller til at skjule eller vise konfigurerbare sektioner i hele appen.",
   "preview_feature_description_leaderboard": "Ranger de personer, du følger, efter set tid, og se hvor du lander blandt dem.",
-  "preview_feature_description_month_in_review": "Forhåndsvis månedsresumé, der nu kører direkte i V3.",
   "preview_feature_description_rewatch": "Start gensynssessioner for serier, du allerede har set, tjek ind eller marker sete episoder direkte fra sessionsskuffen.",
   "preview_feature_description_scoped_favorites": "Tidsbegræns favoritlister ved at vise indeværende års favoritter på profiler og gruppere den fulde favoritliste efter år.",
   "preview_feature_description_smart_recommendations": "Brug vores næste-generations motor til at finde anbefalede film og serier ud over det åbenlyse.",
   "preview_feature_description_smart_related": "Brug vores næste-generations motor til at finde relaterede film og serier ud over det åbenlyse.",
-  "preview_feature_description_year_in_review": "Forhåndsvis årsresumé, der nu kører direkte i V3.",
   "preview_feature_title_edit_mode": "Redigeringstilstand",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Månedsresumé",
   "preview_feature_title_rewatch": "Gensyn",
   "preview_feature_title_scoped_favorites": "Afgrænsede favoritter",
   "preview_feature_title_smart_recommendations": "Smarte anbefalinger",
   "preview_feature_title_smart_related": "Smart relateret",
-  "preview_feature_title_year_in_review": "Årsresumé",
   "reset_status_cleared": "Huskeliste ryddet.",
   "section_title_season_progress": "Sæson {number} status",
   "section_title_seasons_overview": "Oversigt",
