+++ projects/client/i18n/messages/de-DE.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Jahresstatistik",
   "preview_feature_description_edit_mode": "Aktivieren Sie Steuerelemente zum Ausblenden oder Anzeigen konfigurierbarer Abschnitte in der gesamten App.",
   "preview_feature_description_leaderboard": "Rangiere die Personen, denen du folgst, nach Wiedergabezeit und sieh, wo du unter ihnen landest.",
-  "preview_feature_description_month_in_review": "Vorschau auf den Monatsrückblick, jetzt direkt in V3 verfügbar.",
   "preview_feature_description_rewatch": "Starten Sie Wiederholungssitzungen für bereits gesehene Serien, checken Sie Episoden ein oder markieren Sie diese direkt aus dem Sitzungs-Drawer als gesehen.",
   "preview_feature_description_scoped_favorites": "Zeitlich begrenzte Favoritenlisten, indem Favoriten des aktuellen Jahres in Profilen angezeigt und die vollständige Favoritenliste nach Jahr gruppiert werden.",
   "preview_feature_description_smart_recommendations": "Nutze unsere Engine der nächsten Generation, um empfohlene Filme und Serien über das Offensichtliche hinaus aufzudecken.",
   "preview_feature_description_smart_related": "Nutze unsere Engine der nächsten Generation, um verwandte Filme und Serien über das Offensichtliche hinaus aufzudecken.",
-  "preview_feature_description_year_in_review": "Vorschau auf den Jahresrückblick, jetzt direkt in V3 verfügbar.",
   "preview_feature_title_edit_mode": "Bearbeitungsmodus",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Monatsrückblick",
   "preview_feature_title_rewatch": "Wiederholung",
   "preview_feature_title_scoped_favorites": "Gefilterte Favoriten",
   "preview_feature_title_smart_recommendations": "Intelligente Empfehlungen",
   "preview_feature_title_smart_related": "Smarte Verknüpfungen",
-  "preview_feature_title_year_in_review": "Jahresrückblick",
   "reset_status_cleared": "Watchlist geleert.",
   "section_title_season_progress": "Staffel {number} Fortschritt",
   "section_title_seasons_overview": "Übersicht",
