+++ projects/client/i18n/messages/el-GR.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Έτος μέχρι σήμερα",
   "preview_feature_description_edit_mode": "Ενεργοποιήστε τα χειριστήρια για απόκρυψη ή εμφάνιση διαμορφώσιμων ενοτήτων σε όλη την εφαρμογή.",
   "preview_feature_description_leaderboard": "Κατατάξτε τους ανθρώπους που ακολουθείτε με βάση τον χρόνο παρακολούθησης και δείτε πού βρίσκεστε ανάμεσά τους.",
-  "preview_feature_description_month_in_review": "Προεπισκόπηση Μηνιαίας Ανασκόπησης, τώρα διαθέσιμη απευθείας στο V3.",
   "preview_feature_description_rewatch": "Ξεκινήστε συνεδρίες επανάληψης παρακολούθησης για εκπομπές που έχετε ήδη δει, κάντε check-in ή επισημάνετε επεισόδια ως παρακολουθημένα απευθείας από το συρτάρι συνεδριών.",
   "preview_feature_description_scoped_favorites": "Οριοθετήστε τις λίστες αγαπημένων εμφανίζοντας τα αγαπημένα του τρέχοντος έτους στα προφίλ και ομαδοποιώντας την πλήρη λίστα αγαπημένων ανά έτος.",
   "preview_feature_description_smart_recommendations": "Χρησιμοποιήστε τον κινητήρα επόμενης γενιάς μας για να ανακαλύψετε προτεινόμενες ταινίες και σειρές πέρα από τα προφανή.",
   "preview_feature_description_smart_related": "Χρησιμοποιήστε τον κινητήρα επόμενης γενιάς μας για να ανακαλύψετε σχετικές ταινίες και σειρές πέρα από τα προφανή.",
-  "preview_feature_description_year_in_review": "Προεπισκόπηση Ετήσιας Ανασκόπησης, τώρα διαθέσιμη απευθείας στο V3.",
   "preview_feature_title_edit_mode": "Λειτουργία επεξεργασίας",
   "preview_feature_title_leaderboard": "Πάνθεον",
-  "preview_feature_title_month_in_review": "Μηνιαία Ανασκόπηση",
   "preview_feature_title_rewatch": "Επανάληψη",
   "preview_feature_title_scoped_favorites": "Αγαπημένα με χρονικό πλαίσιο",
   "preview_feature_title_smart_recommendations": "Έξυπνες Προτάσεις",
   "preview_feature_title_smart_related": "Έξυπνα Σχετικά",
-  "preview_feature_title_year_in_review": "Ετήσια Ανασκόπηση",
   "reset_status_cleared": "Η λίστα παρακολούθησης εκκαθαρίστηκε.",
   "section_title_season_progress": "Πρόοδος Σεζόν {number}",
   "section_title_seasons_overview": "Επισκόπηση",
