+++ projects/client/i18n/messages/en-AU.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Year to date",
   "preview_feature_description_edit_mode": "Enable controls for hiding or showing configurable sections across the app.",
   "preview_feature_description_leaderboard": "Rank the people you follow by watch time and see where you land among them.",
-  "preview_feature_description_month_in_review": "Preview Month in Review, now running directly in V3.",
   "preview_feature_description_rewatch": "Start rewatching sessions for shows you've already seen, check in or mark watched episodes directly from the session drawer.",
   "preview_feature_description_scoped_favorites": "Time-box favourite lists by showing current-year favourites on profiles and grouping the full favourites list by year.",
   "preview_feature_description_smart_recommendations": "Use our next-gen engine to surface recommended movies and shows beyond the obvious.",
   "preview_feature_description_smart_related": "Use our next-gen engine to surface related movies and shows beyond the obvious.",
-  "preview_feature_description_year_in_review": "Preview Year in Review, now running directly in V3.",
   "preview_feature_title_edit_mode": "Edit Mode",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Month In Review",
   "preview_feature_title_rewatch": "Rewatch",
   "preview_feature_title_scoped_favorites": "Scoped Favourites",
   "preview_feature_title_smart_recommendations": "Smart Recommendations",
   "preview_feature_title_smart_related": "Smart Related",
-  "preview_feature_title_year_in_review": "Year In Review",
   "reset_status_cleared": "Watchlist cleared.",
   "section_title_season_progress": "Season {number} progress",
   "section_title_seasons_overview": "Overview",
