+++ projects/client/i18n/messages/es-ES.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "En lo que va del año",
   "preview_feature_description_edit_mode": "Habilita los controles para ocultar o mostrar secciones configurables en toda la aplicación.",
   "preview_feature_description_leaderboard": "Clasifica a las personas que sigues por tiempo de visualización y descubre tu posición entre ellas.",
-  "preview_feature_description_month_in_review": "Vista previa de Tu mes en Trakt, ahora ejecutándose directamente en V3.",
   "preview_feature_description_rewatch": "Inicia sesiones de re-visualización para programas que ya has visto, registra o marca los episodios como vistos directamente desde el cajón de la sesión.",
   "preview_feature_description_scoped_favorites": "Limita las listas de favoritos por tiempo mostrando los favoritos del año actual en los perfiles y agrupando la lista completa de favoritos por año.",
   "preview_feature_description_smart_recommendations": "Utiliza nuestro motor de próxima generación para descubrir películas y series recomendadas más allá de lo obvio.",
   "preview_feature_description_smart_related": "Utiliza nuestro motor de próxima generación para descubrir películas y series relacionadas más allá de lo obvio.",
-  "preview_feature_description_year_in_review": "Vista previa de Tu año en Trakt, ahora ejecutándose directamente en V3.",
   "preview_feature_title_edit_mode": "Modo de edición",
   "preview_feature_title_leaderboard": "Panteón",
-  "preview_feature_title_month_in_review": "Tu mes en Trakt",
   "preview_feature_title_rewatch": "Re-visualización",
   "preview_feature_title_scoped_favorites": "Favoritos delimitados",
   "preview_feature_title_smart_recommendations": "Recomendaciones inteligentes",
   "preview_feature_title_smart_related": "Relacionados Inteligentes",
-  "preview_feature_title_year_in_review": "Tu año en Trakt",
   "reset_status_cleared": "Lista de seguimiento borrada.",
   "section_title_season_progress": "Temporada {number} progreso",
   "section_title_seasons_overview": "Resumen",
