+++ projects/client/i18n/messages/es-MX.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Año hasta la fecha",
   "preview_feature_description_edit_mode": "Habilita los controles para ocultar o mostrar secciones configurables en toda la aplicación.",
   "preview_feature_description_leaderboard": "Clasifica a las personas que sigues por tiempo de visualización y ve dónde te encuentras entre ellas.",
-  "preview_feature_description_month_in_review": "Previsualiza tu Resumen Mensual, ahora ejecutándose directamente en la V3.",
   "preview_feature_description_rewatch": "Inicia sesiones para volver a ver programas que ya has visto, haz check-in o marca episodios como vistos directamente desde el panel de sesión.",
   "preview_feature_description_scoped_favorites": "Delimita las listas de favoritos mostrando los favoritos del año actual en los perfiles y agrupando la lista completa de favoritos por año.",
   "preview_feature_description_smart_recommendations": "Usa nuestro motor de próxima generación para encontrar películas y programas recomendados más allá de lo obvio.",
   "preview_feature_description_smart_related": "Usa nuestro motor de próxima generación para encontrar películas y programas relacionados más allá de lo obvio.",
-  "preview_feature_description_year_in_review": "Previsualiza tu Resumen Anual, ahora ejecutándose directamente en la V3.",
   "preview_feature_title_edit_mode": "Modo de edición",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Resumen Mensual",
   "preview_feature_title_rewatch": "Volver a ver",
   "preview_feature_title_scoped_favorites": "Favoritos con Alcance",
   "preview_feature_title_smart_recommendations": "Recomendaciones inteligentes",
   "preview_feature_title_smart_related": "Relacionado Inteligente",
-  "preview_feature_title_year_in_review": "Resumen Anual",
   "reset_status_cleared": "Lista de seguimiento borrada.",
   "section_title_season_progress": "Progreso de la temporada {number}",
   "section_title_seasons_overview": "Resumen",
