+++ projects/client/i18n/messages/fa-IR.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "سال تا به امروز",
   "preview_feature_description_edit_mode": "کنترل‌هایی را برای پنهان کردن یا نمایش بخش‌های قابل تنظیم در سراسر برنامه فعال کنید.",
   "preview_feature_description_leaderboard": "افرادی که دنبال می‌کنید را بر اساس زمان تماشا رتبه‌بندی کنید و ببینید در میان آن‌ها کجا قرار می‌گیرید.",
-  "preview_feature_description_month_in_review": "مرور ماه را پیش‌نمایش کنید، اکنون مستقیماً در V3 اجرا می‌شود.",
   "preview_feature_description_rewatch": "جلسات بازتماشا را برای سریال‌هایی که قبلاً دیده‌اید شروع کنید، قسمت‌های تماشا شده را مستقیماً از کشوی جلسه ثبت کنید یا علامت بزنید.",
   "preview_feature_description_scoped_favorites": "لیست‌های علاقه‌مندی را با نمایش علاقه‌مندی‌های سال جاری در پروفایل‌ها و گروه‌بندی لیست کامل علاقه‌مندی‌ها بر اساس سال، محدود کنید.",
   "preview_feature_description_smart_recommendations": "از موتور نسل بعدی ما برای نمایش فیلم‌ها و نمایش‌های توصیه شده فراتر از موارد بدیهی استفاده کنید.",
   "preview_feature_description_smart_related": "از موتور نسل بعدی ما برای نمایش فیلم‌ها و نمایش‌های مرتبط فراتر از موارد بدیهی استفاده کنید.",
-  "preview_feature_description_year_in_review": "مرور سال را پیش‌نمایش کنید، اکنون مستقیماً در V3 اجرا می‌شود.",
   "preview_feature_title_edit_mode": "حالت ویرایش",
   "preview_feature_title_leaderboard": "پانتئون",
-  "preview_feature_title_month_in_review": "مرور ماه",
   "preview_feature_title_rewatch": "بازتماشا",
   "preview_feature_title_scoped_favorites": "علاقه‌مندی‌های محدود",
   "preview_feature_title_smart_recommendations": "توصیه‌های هوشمند",
   "preview_feature_title_smart_related": "مرتبط هوشمند",
-  "preview_feature_title_year_in_review": "مرور سال",
   "reset_status_cleared": "لیست تماشا پاک شد.",
   "section_title_season_progress": "فصل {number} پیشرفت",
   "section_title_seasons_overview": "مرور کلی",
