+++ projects/client/i18n/messages/fr-FR.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Année à ce jour",
   "preview_feature_description_edit_mode": "Activez les commandes pour masquer ou afficher les sections configurables dans toute l'application.",
   "preview_feature_description_leaderboard": "Classez les personnes que vous suivez par temps de visionnage et découvrez votre place parmi elles.",
-  "preview_feature_description_month_in_review": "Aperçu du bilan mensuel, désormais exécuté directement dans la V3.",
   "preview_feature_description_rewatch": "Démarrez des sessions de revisionnage pour les séries que vous avez déjà vues, enregistrez-vous ou marquez les épisodes comme vus directement depuis le tiroir de session.",
   "preview_feature_description_scoped_favorites": "Délimitez les listes de favoris en affichant les favoris de l'année en cours sur les profils et en regroupant la liste complète des favoris par année.",
   "preview_feature_description_smart_recommendations": "Utilisez notre moteur de nouvelle génération pour découvrir des films et séries recommandés au-delà de l'évidence.",
   "preview_feature_description_smart_related": "Utilisez notre moteur de nouvelle génération pour découvrir des films et séries associés au-delà de l'évidence.",
-  "preview_feature_description_year_in_review": "Aperçu du bilan annuel, désormais exécuté directement dans la V3.",
   "preview_feature_title_edit_mode": "Mode édition",
   "preview_feature_title_leaderboard": "Panthéon",
-  "preview_feature_title_month_in_review": "Bilan mensuel",
   "preview_feature_title_rewatch": "Revisionnage",
   "preview_feature_title_scoped_favorites": "Favoris délimités",
   "preview_feature_title_smart_recommendations": "Recommandations intelligentes",
   "preview_feature_title_smart_related": "Contenu associé intelligent",
-  "preview_feature_title_year_in_review": "Bilan annuel",
   "reset_status_cleared": "Liste de suivi vidée.",
   "section_title_season_progress": "Saison {number} progression",
   "section_title_seasons_overview": "Aperçu",
