+++ projects/client/i18n/messages/hu-HU.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Éves összesítés",
   "preview_feature_description_edit_mode": "Engedélyezze a vezérlőket a konfigurálható szakaszok elrejtéséhez vagy megjelenítéséhez az alkalmazásban.",
   "preview_feature_description_leaderboard": "Rangsorold a követett embereket a nézési idő alapján, és nézd meg, hol helyezkedsz el közöttük.",
-  "preview_feature_description_month_in_review": "Előzetes Havi összefoglaló, mostantól közvetlenül a V3-ban fut.",
   "preview_feature_description_rewatch": "Indíts újra nézési munkameneteket olyan műsorokhoz, amiket már láttál, jelentkezz be, vagy jelölj meg nézett epizódokat közvetlenül a munkamenet fiókból.",
   "preview_feature_description_scoped_favorites": "Időkorlátozza a kedvenc listákat azáltal, hogy megjeleníti az aktuális év kedvenceit a profilokon, és csoportosítja a teljes kedvencek listáját év szerint.",
   "preview_feature_description_smart_recommendations": "Használd következő generációs motorunkat, hogy a nyilvánvalón túlmutató ajánlott filmeket és sorozatokat fedezz fel.",
   "preview_feature_description_smart_related": "Használd következő generációs motorunkat, hogy a nyilvánvalón túlmutató kapcsolódó filmeket és sorozatokat fedezz fel.",
-  "preview_feature_description_year_in_review": "Előzetes Éves összefoglaló, mostantól közvetlenül a V3-ban fut.",
   "preview_feature_title_edit_mode": "Szerkesztési mód",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Havi összefoglaló",
   "preview_feature_title_rewatch": "Újra nézés",
   "preview_feature_title_scoped_favorites": "Korlátozott Kedvencek",
   "preview_feature_title_smart_recommendations": "Intelligens ajánlások",
   "preview_feature_title_smart_related": "Okos Kapcsolódó",
-  "preview_feature_title_year_in_review": "Éves összefoglaló",
   "reset_status_cleared": "Megnézendők listája kiürítve.",
   "section_title_season_progress": "{number}. évad előrehaladása",
   "section_title_seasons_overview": "Áttekintés",
