+++ projects/client/i18n/messages/id-ID.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Tahun Berjalan",
   "preview_feature_description_edit_mode": "Aktifkan kontrol untuk menyembunyikan atau menampilkan bagian yang dapat dikonfigurasi di seluruh aplikasi.",
   "preview_feature_description_leaderboard": "Peringkat orang yang Anda ikuti berdasarkan waktu tonton dan lihat posisi Anda di antara mereka.",
-  "preview_feature_description_month_in_review": "Pratinjau Kilas Balik Bulanan, sekarang berjalan langsung di V3.",
   "preview_feature_description_rewatch": "Mulai sesi menonton ulang untuk acara yang sudah Anda tonton, check-in atau tandai episode yang sudah ditonton langsung dari panel sesi.",
   "preview_feature_description_scoped_favorites": "Batas waktu daftar favorit dengan menampilkan favorit tahun berjalan di profil dan mengelompokkan daftar favorit lengkap berdasarkan tahun.",
   "preview_feature_description_smart_recommendations": "Gunakan mesin generasi berikutnya kami untuk menampilkan film dan acara yang direkomendasikan melampaui yang biasa.",
   "preview_feature_description_smart_related": "Gunakan mesin generasi berikutnya kami untuk menampilkan film dan acara terkait yang melampaui yang biasa.",
-  "preview_feature_description_year_in_review": "Pratinjau Kilas Balik Tahunan, sekarang berjalan langsung di V3.",
   "preview_feature_title_edit_mode": "Mode Edit",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Kilas Balik Bulanan",
   "preview_feature_title_rewatch": "Tonton Ulang",
   "preview_feature_title_scoped_favorites": "Favorit Terbatas",
   "preview_feature_title_smart_recommendations": "Rekomendasi Cerdas",
   "preview_feature_title_smart_related": "Terkait Cerdas",
-  "preview_feature_title_year_in_review": "Kilas Balik Tahunan",
   "reset_status_cleared": "Daftar tontonan dibersihkan.",
   "section_title_season_progress": "Musim {number} kemajuan",
   "section_title_seasons_overview": "Ikhtisar",
