+++ projects/client/i18n/messages/it-IT.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Dall'inizio dell'anno",
   "preview_feature_description_edit_mode": "Abilita i controlli per nascondere o mostrare sezioni configurabili in tutta l'app.",
   "preview_feature_description_leaderboard": "Classifica le persone che segui in base al tempo di visione e scopri dove ti posizioni tra loro.",
-  "preview_feature_description_month_in_review": "Anteprima Riepilogo del mese, ora in esecuzione direttamente in V3.",
   "preview_feature_description_rewatch": "Avvia sessioni di riguardato per le serie TV che hai già visto, registrando la visione o contrassegnando gli episodi come guardati direttamente dal pannello della sessione.",
   "preview_feature_description_scoped_favorites": "Limita le liste dei preferiti per anno, mostrando i preferiti dell'anno corrente sui profili e raggruppando l'intera lista dei preferiti per anno.",
   "preview_feature_description_smart_recommendations": "Utilizza il nostro motore di nuova generazione per scoprire film e serie TV raccomandati che vanno oltre l'ovvio.",
   "preview_feature_description_smart_related": "Utilizza il nostro motore di nuova generazione per scoprire film e serie TV correlati che vanno oltre l'ovvio.",
-  "preview_feature_description_year_in_review": "Anteprima Riepilogo dell'anno, ora in esecuzione direttamente in V3.",
   "preview_feature_title_edit_mode": "Modalità di modifica",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Riepilogo del mese",
   "preview_feature_title_rewatch": "Riguarda",
   "preview_feature_title_scoped_favorites": "Preferiti con Ambito",
   "preview_feature_title_smart_recommendations": "Raccomandazioni Smart",
   "preview_feature_title_smart_related": "Correlati Intelligenti",
-  "preview_feature_title_year_in_review": "Riepilogo dell'anno",
   "reset_status_cleared": "Watchlist svuotata.",
   "section_title_season_progress": "Stagione {number} progressi",
   "section_title_seasons_overview": "Panoramica",
