+++ projects/client/i18n/messages/ja-JP.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "今年これまでの統計",
   "preview_feature_description_edit_mode": "アプリ全体で設定可能なセクションの表示/非表示を制御する機能を有効にします。",
   "preview_feature_description_leaderboard": "フォローしている人の視聴時間でランキングを作成し、その中で自分がどこに位置するかを確認できます。",
-  "preview_feature_description_month_in_review": "V3で直接実行される月間レビューをプレビューします。",
   "preview_feature_description_rewatch": "既に視聴済みの番組の再視聴セッションを開始し、セッションドロワーから直接エピソードをチェックインまたは視聴済みとしてマークできます。",
   "preview_feature_description_scoped_favorites": "プロフィールに今年の「お気に入り」を表示し、お気に入りリスト全体を年別にグループ化することで、お気に入りリストを期間で区切ります。",
   "preview_feature_description_smart_recommendations": "当社の次世代エンジンを活用して、一般的なおすすめを超えた映画や番組を発見しましょう。",
   "preview_feature_description_smart_related": "当社の次世代エンジンを活用して、一般的な関連作を超えた映画や番組を発見しましょう。",
-  "preview_feature_description_year_in_review": "V3で直接実行される年間レビューをプレビューします。",
   "preview_feature_title_edit_mode": "編集モード",
   "preview_feature_title_leaderboard": "パンテオン",
-  "preview_feature_title_month_in_review": "月間レビュー",
   "preview_feature_title_rewatch": "再視聴",
   "preview_feature_title_scoped_favorites": "範囲指定されたお気に入り",
   "preview_feature_title_smart_recommendations": "スマートレコメンデーション",
   "preview_feature_title_smart_related": "スマート関連",
-  "preview_feature_title_year_in_review": "年間レビュー",
   "reset_status_cleared": "ウォッチリストをクリアしました。",
   "section_title_season_progress": "シーズン{number}の進捗",
   "section_title_seasons_overview": "概要",
