+++ projects/client/i18n/messages/nb-NO.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "År til dato",
   "preview_feature_description_edit_mode": "Aktiver kontroller for å skjule eller vise konfigurerbare seksjoner i hele appen.",
   "preview_feature_description_leaderboard": "Ranger personene du følger etter seertid og se hvor du lander blant dem.",
-  "preview_feature_description_month_in_review": "Forhåndsvis Måneden i retrospekt, kjører nå direkte i V3.",
   "preview_feature_description_rewatch": "Start gjensynsøkter for serier du allerede har sett, sjekk inn eller merk episoder som sett direkte fra gjensynsskuffen.",
   "preview_feature_description_scoped_favorites": "Tidsbegrens favorittlister ved å vise inneværende års favoritter på profiler og gruppere den komplette favorittlisten etter år.",
   "preview_feature_description_smart_recommendations": "Bruk vår neste generasjons motor til å avdekke anbefalte filmer og serier utover det åpenbare.",
   "preview_feature_description_smart_related": "Bruk vår neste generasjons motor til å avdekke relaterte filmer og serier utover det åpenbare.",
-  "preview_feature_description_year_in_review": "Forhåndsvis Året i retrospekt, kjører nå direkte i V3.",
   "preview_feature_title_edit_mode": "Redigeringsmodus",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Måneden i retrospekt",
   "preview_feature_title_rewatch": "Gjensyn",
   "preview_feature_title_scoped_favorites": "Avgrensede favoritter",
   "preview_feature_title_smart_recommendations": "Smarte anbefalinger",
   "preview_feature_title_smart_related": "Smarte Relaterte",
-  "preview_feature_title_year_in_review": "Året i retrospekt",
   "reset_status_cleared": "Overvåkningsliste tømt.",
   "section_title_season_progress": "Sesong {number} fremdrift",
   "section_title_seasons_overview": "Oversikt",
