+++ projects/client/i18n/messages/nl-NL.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Jaar tot nu toe",
   "preview_feature_description_edit_mode": "Schakel besturingselementen in voor het verbergen of weergeven van configureerbare secties in de app.",
   "preview_feature_description_leaderboard": "Rangschik de mensen die je volgt op kijktijd en zie waar jij eindigt.",
-  "preview_feature_description_month_in_review": "Bekijk Maandoverzicht als preview, nu direct beschikbaar in V3.",
   "preview_feature_description_rewatch": "Start nieuwe kijkbeurten voor series die je al hebt gezien, check in of markeer afleveringen als bekeken, direct vanuit het sessiepaneel.",
   "preview_feature_description_scoped_favorites": "Beperk favorietenlijsten door favorieten van het huidige jaar op profielen weer te geven en de volledige favorietenlijst per jaar te groeperen.",
   "preview_feature_description_smart_recommendations": "Gebruik onze next-gen engine om aanbevolen films en series te vinden die verder gaan dan het voor de hand liggende.",
   "preview_feature_description_smart_related": "Gebruik onze next-gen engine om gerelateerde films en series te vinden die verder gaan dan het voor de hand liggende.",
-  "preview_feature_description_year_in_review": "Bekijk Jaaroverzicht als preview, nu direct beschikbaar in V3.",
   "preview_feature_title_edit_mode": "Bewerkmodus",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Maandoverzicht",
   "preview_feature_title_rewatch": "Opnieuw bekijken",
   "preview_feature_title_scoped_favorites": "Gerichte favorieten",
   "preview_feature_title_smart_recommendations": "Slimme aanbevelingen",
   "preview_feature_title_smart_related": "Slim Gerelateerd",
-  "preview_feature_title_year_in_review": "Jaaroverzicht",
   "reset_status_cleared": "Kijklijst gewist.",
   "section_title_season_progress": "Seizoen {number} voortgang",
   "section_title_seasons_overview": "Overzicht",
