+++ projects/client/i18n/messages/pl-PL.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Od początku roku",
   "preview_feature_description_edit_mode": "Włącz sterowanie ukrywaniem lub pokazywaniem konfigurowalnych sekcji w całej aplikacji.",
   "preview_feature_description_leaderboard": "Oceniaj osoby, które obserwujesz, według czasu oglądania i sprawdź, gdzie znajdziesz się wśród nich.",
-  "preview_feature_description_month_in_review": "Podgląd Podsumowania miesiąca, teraz działający bezpośrednio w V3.",
   "preview_feature_description_rewatch": "Rozpocznij sesje ponownego oglądania dla programów, które już widziałeś, melduj się lub oznaczaj obejrzane odcinki bezpośrednio z szuflady sesji.",
   "preview_feature_description_scoped_favorites": "Ograniczaj listy ulubionych czasowo, pokazując ulubione z bieżącego roku w profilach i grupując pełną listę ulubionych według roku.",
   "preview_feature_description_smart_recommendations": "Użyj naszego silnika nowej generacji, aby odkryć polecane filmy i programy wykraczające poza oczywiste.",
   "preview_feature_description_smart_related": "Użyj naszego silnika nowej generacji, aby odkryć powiązane filmy i programy wykraczające poza oczywiste.",
-  "preview_feature_description_year_in_review": "Podgląd Podsumowania roku, teraz działający bezpośrednio w V3.",
   "preview_feature_title_edit_mode": "Tryb edycji",
   "preview_feature_title_leaderboard": "Panteon",
-  "preview_feature_title_month_in_review": "Podsumowanie miesiąca",
   "preview_feature_title_rewatch": "Ponowne oglądanie",
   "preview_feature_title_scoped_favorites": "Ulubione z podziałem",
   "preview_feature_title_smart_recommendations": "Inteligentne rekomendacje",
   "preview_feature_title_smart_related": "Inteligentne Powiązane",
-  "preview_feature_title_year_in_review": "Podsumowanie roku",
   "reset_status_cleared": "Lista obserwowanych usunięta.",
   "section_title_season_progress": "Sezon {number} postęp",
   "section_title_seasons_overview": "Przegląd",
