+++ projects/client/i18n/messages/pt-BR.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Acumulado do Ano",
   "preview_feature_description_edit_mode": "Ative os controles para ocultar ou exibir seções configuráveis em todo o aplicativo.",
   "preview_feature_description_leaderboard": "Classifique as pessoas que você segue por tempo de exibição e veja sua posição entre elas.",
-  "preview_feature_description_month_in_review": "Pré-visualize a Retrospectiva do Mês, agora rodando diretamente na V3.",
   "preview_feature_description_rewatch": "Comece sessões de reassistência para séries que você já assistiu, faça check-in ou marque episódios como assistidos diretamente na gaveta da sessão.",
   "preview_feature_description_scoped_favorites": "Limite as listas de favoritos por tempo, mostrando os favoritos do ano atual nos perfis e agrupando a lista completa de favoritos por ano.",
   "preview_feature_description_smart_recommendations": "Use nosso mecanismo de próxima geração para descobrir filmes e séries recomendados, além do óbvio.",
   "preview_feature_description_smart_related": "Use nosso mecanismo de próxima geração para descobrir filmes e séries relacionados, além do óbvio.",
-  "preview_feature_description_year_in_review": "Pré-visualize a Retrospectiva do Ano, agora rodando diretamente na V3.",
   "preview_feature_title_edit_mode": "Modo de edição",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Retrospectiva do Mês",
   "preview_feature_title_rewatch": "Reassistir",
   "preview_feature_title_scoped_favorites": "Favoritos Restritos",
   "preview_feature_title_smart_recommendations": "Recomendações Inteligentes",
   "preview_feature_title_smart_related": "Relacionados Inteligentes",
-  "preview_feature_title_year_in_review": "Retrospectiva do Ano",
   "reset_status_cleared": "Lista para Assistir limpa.",
   "section_title_season_progress": "Progresso da temporada {number}",
   "section_title_seasons_overview": "Visão Geral",
