+++ projects/client/i18n/messages/pt-PT.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Ano até ao momento",
   "preview_feature_description_edit_mode": "Ativa controlos para mostrar ou ocultar secções configuráveis em toda a app.",
   "preview_feature_description_leaderboard": "Ordena as pessoas que segues pelo tempo de visualização e vê em que posição ficas entre elas.",
-  "preview_feature_description_month_in_review": "Experimenta o Resumo do mês, agora integrado diretamente na V3.",
   "preview_feature_description_rewatch": "Inicia sessões de revisão para séries que já viste e regista a visualização ou marca episódios como vistos diretamente no painel da sessão.",
   "preview_feature_description_scoped_favorites": "Organiza a lista de favoritos mostrando os do ano atual no perfil e agrupando a lista completa por ano.",
   "preview_feature_description_smart_recommendations": "Utiliza o nosso motor de recomendação de nova geração para descobrir filmes e séries para além das recomendações mais óbvias.",
   "preview_feature_description_smart_related": "Utiliza o nosso motor de nova geração para descobrires recomendações personalizadas de filmes e séries que vão além do óbvio.",
-  "preview_feature_description_year_in_review": "Experimenta o Resumo do ano, agora integrado diretamente na V3.",
   "preview_feature_title_edit_mode": "Modo de edição",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Resumo do mês",
   "preview_feature_title_rewatch": "Rever",
   "preview_feature_title_scoped_favorites": "Favoritos por ano",
   "preview_feature_title_smart_recommendations": "Recomendações inteligentes",
   "preview_feature_title_smart_related": "Recomendações inteligentes",
-  "preview_feature_title_year_in_review": "Resumo do ano",
   "reset_status_cleared": "Lista para ver limpa.",
   "section_title_season_progress": "Progresso da temporada {number}",
   "section_title_seasons_overview": "Sinopse",
