+++ projects/client/i18n/messages/ro-RO.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Anul curent",
   "preview_feature_description_edit_mode": "Activează controale pentru ascunderea sau afișarea secțiunilor configurabile în întreaga aplicație.",
   "preview_feature_description_leaderboard": "Clasați persoanele pe care le urmăriți după timpul de vizionare și vedeți unde vă situați printre ele.",
-  "preview_feature_description_month_in_review": "Previzualizează Retrospectiva Lunii, acum disponibilă direct în V3.",
   "preview_feature_description_rewatch": "Începeți sesiuni de revizionare pentru emisiunile pe care le-ați văzut deja, înregistrați-vă sau marcați episoadele vizionate direct din panoul de sesiune.",
   "preview_feature_description_scoped_favorites": "Organizează listele de favorite pe perioade, afișând favoritele din anul curent pe profile și grupând lista completă de favorite pe ani.",
   "preview_feature_description_smart_recommendations": "Utilizează motorul nostru de ultimă generație pentru a descoperi filme și seriale recomandate, dincolo de cele evidente.",
   "preview_feature_description_smart_related": "Utilizează motorul nostru de ultimă generație pentru a descoperi filme și seriale corelate, dincolo de cele evidente.",
-  "preview_feature_description_year_in_review": "Previzualizează Retrospectiva Anului, acum disponibilă direct în V3.",
   "preview_feature_title_edit_mode": "Mod editare",
   "preview_feature_title_leaderboard": "Panteon",
-  "preview_feature_title_month_in_review": "Retrospectiva Lunii",
   "preview_feature_title_rewatch": "Revizionare",
   "preview_feature_title_scoped_favorites": "Favorite Contextuale",
   "preview_feature_title_smart_recommendations": "Recomandări inteligente",
   "preview_feature_title_smart_related": "Inteligent Corelate",
-  "preview_feature_title_year_in_review": "Retrospectiva Anului",
   "reset_status_cleared": "Lista de vizionare a fost ștearsă.",
   "section_title_season_progress": "Sezonul {number} progres",
   "section_title_seasons_overview": "Prezentare generală",
