+++ projects/client/i18n/messages/ru-RU.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "С начала года",
   "preview_feature_description_edit_mode": "Включите элементы управления для скрытия или отображения настраиваемых разделов во всем приложении.",
   "preview_feature_description_leaderboard": "Оценивайте людей, на которых вы подписаны, по времени просмотра и смотрите, какое место вы занимаете среди них.",
-  "preview_feature_description_month_in_review": "Предварительный просмотр «Обзора месяца», теперь работающий непосредственно в V3.",
   "preview_feature_description_rewatch": "Начните сеансы пересмотра для шоу, которые вы уже видели, отмечайте просмотр или помечайте эпизоды как просмотренные прямо из панели сеанса.",
   "preview_feature_description_scoped_favorites": "Ограничивайте списки избранного по времени, показывая избранное текущего года в профилях и группируя полный список избранного по годам.",
   "preview_feature_description_smart_recommendations": "Используйте наш движок нового поколения, чтобы находить рекомендуемые фильмы и шоу, выходящие за рамки очевидного.",
   "preview_feature_description_smart_related": "Используйте наш движок нового поколения, чтобы находить связанные фильмы и шоу, выходящие за рамки очевидного.",
-  "preview_feature_description_year_in_review": "Предварительный просмотр «Обзора года», теперь работающий непосредственно в V3.",
   "preview_feature_title_edit_mode": "Режим редактирования",
   "preview_feature_title_leaderboard": "Пантеон",
-  "preview_feature_title_month_in_review": "Обзор месяца",
   "preview_feature_title_rewatch": "Пересмотр",
   "preview_feature_title_scoped_favorites": "Избранное по времени",
   "preview_feature_title_smart_recommendations": "Умные рекомендации",
   "preview_feature_title_smart_related": "Умные рекомендации",
-  "preview_feature_title_year_in_review": "Обзор года",
   "reset_status_cleared": "Список просмотра очищен.",
   "section_title_season_progress": "Прогресс сезона {number}",
   "section_title_seasons_overview": "Обзор",
