+++ projects/client/i18n/messages/sv-SE.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "År hittills",
   "preview_feature_description_edit_mode": "Aktivera kontroller för att dölja eller visa konfigurerbara sektioner i hela appen.",
   "preview_feature_description_leaderboard": "Rankar de personer du följer efter tittartid och se var du hamnar bland dem.",
-  "preview_feature_description_month_in_review": "Förhandsgranska månadsöversikten, nu körs direkt i V3.",
   "preview_feature_description_rewatch": "Starta omtittningssessioner för serier du redan har sett, checka in eller markera avsnitt som sedda direkt från sessionslådan.",
   "preview_feature_description_scoped_favorites": "Tidsavgränsa favoritlistor genom att visa favoriter från innevarande år på profiler och gruppera hela favoritlistan per år.",
   "preview_feature_description_smart_recommendations": "Använd vår nästa generations motor för att upptäcka rekommenderade filmer och serier bortom det uppenbara.",
   "preview_feature_description_smart_related": "Använd vår nästa generations motor för att upptäcka relaterade filmer och serier bortom det uppenbara.",
-  "preview_feature_description_year_in_review": "Förhandsgranska årsöversikten, nu körs direkt i V3.",
   "preview_feature_title_edit_mode": "Redigeringsläge",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Månadsöversikt",
   "preview_feature_title_rewatch": "Omtittning",
   "preview_feature_title_scoped_favorites": "Avgränsade favoriter",
   "preview_feature_title_smart_recommendations": "Smarta rekommendationer",
   "preview_feature_title_smart_related": "Smart Relaterade",
-  "preview_feature_title_year_in_review": "Årsöversikt",
   "reset_status_cleared": "Bevakningslistan rensad.",
   "section_title_season_progress": "Säsong {number} förlopp",
   "section_title_seasons_overview": "Översikt",
