+++ projects/client/i18n/messages/tr-TR.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Yıl Başından Bugüne",
   "preview_feature_description_edit_mode": "Uygulama genelinde yapılandırılabilir bölümleri gizleme veya gösterme kontrollerini etkinleştirin.",
   "preview_feature_description_leaderboard": "Takip ettiğin kişileri izleme süresine göre sırala ve aralarında nerede olduğunu gör.",
-  "preview_feature_description_month_in_review": "Ayın İncelemesini önizleyin, şimdi doğrudan V3'te çalışıyor.",
   "preview_feature_description_rewatch": "Halihazırda izlediğiniz diziler için yeniden izleme oturumları başlatın, izlenen bölümleri doğrudan oturum çekmecesinden kaydedin veya izlendi olarak işaretleyin.",
   "preview_feature_description_scoped_favorites": "Favori listelerini, profillerde güncel yılın favorilerini göstererek ve tüm favori listesini yıla göre gruplandırarak zaman kutucuklarına ayırın.",
   "preview_feature_description_smart_recommendations": "Sıradanlığın ötesinde önerilen film ve dizileri keşfetmek için yeni nesil motorumuzu kullanın.",
   "preview_feature_description_smart_related": "Sıradanlığın ötesinde ilgili film ve dizileri keşfetmek için yeni nesil motorumuzu kullanın.",
-  "preview_feature_description_year_in_review": "Yılın İncelemesini önizleyin, şimdi doğrudan V3'te çalışıyor.",
   "preview_feature_title_edit_mode": "Düzenleme Modu",
   "preview_feature_title_leaderboard": "Pantheon",
-  "preview_feature_title_month_in_review": "Ayın İncelemesi",
   "preview_feature_title_rewatch": "Yeniden İzleme",
   "preview_feature_title_scoped_favorites": "Kapsamlı Favoriler",
   "preview_feature_title_smart_recommendations": "Akıllı Öneriler",
   "preview_feature_title_smart_related": "Akıllı İlişkili",
-  "preview_feature_title_year_in_review": "Yılın İncelemesi",
   "reset_status_cleared": "İzleme listesi temizlendi.",
   "section_title_season_progress": "Sezon {number} ilerlemesi",
   "section_title_seasons_overview": "Genel Bakış",
