+++ projects/client/i18n/messages/uk-UA.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "Рік до сьогодні",
   "preview_feature_description_edit_mode": "Увімкніть елементи керування для приховування або відображення настроюваних розділів у всьому застосунку.",
   "preview_feature_description_leaderboard": "Ранжуйте людей, за якими стежите, за часом перегляду та подивіться, яке місце ви посідаєте серед них.",
-  "preview_feature_description_month_in_review": "Попередній перегляд Місяць у Підсумках, тепер працює безпосередньо у V3.",
   "preview_feature_description_rewatch": "Почніть сеанси повторного перегляду для шоу, які ви вже бачили, реєструйте або відмічайте переглянуті серії безпосередньо з вікна сеансу.",
   "preview_feature_description_scoped_favorites": "Обмежте списки улюблених за часом, відображаючи улюблене поточного року в профілях та групуючи повний список улюблених за роками.",
   "preview_feature_description_smart_recommendations": "Використовуйте наш двигун нового покоління, щоб знаходити рекомендовані фільми та шоу, що виходять за рамки очевидного.",
   "preview_feature_description_smart_related": "Використовуйте наш двигун нового покоління, щоб знаходити пов'язані фільми та шоу, що виходять за рамки очевидного.",
-  "preview_feature_description_year_in_review": "Попередній перегляд Рік у Підсумках, тепер працює безпосередньо у V3.",
   "preview_feature_title_edit_mode": "Режим редагування",
   "preview_feature_title_leaderboard": "Пантеон",
-  "preview_feature_title_month_in_review": "Місяць у Підсумках",
   "preview_feature_title_rewatch": "Повторний перегляд",
   "preview_feature_title_scoped_favorites": "Вибране за періодом",
   "preview_feature_title_smart_recommendations": "Розумні рекомендації",
   "preview_feature_title_smart_related": "Розумні пов'язані",
-  "preview_feature_title_year_in_review": "Рік у Підсумках",
   "reset_status_cleared": "Список перегляду очищено.",
   "section_title_season_progress": "Прогрес сезону {number}",
   "section_title_seasons_overview": "Огляд",
