+++ projects/client/i18n/messages/zh-CN.json
@@ -1125,20 +1125,16 @@
   "page_title_year_to_date": "年至今",
   "preview_feature_description_edit_mode": "启用控件，用于隐藏或显示应用程序中可配置的部分。",
   "preview_feature_description_leaderboard": "根据观看时长对您关注的人进行排名，查看您在他们中的位置。",
-  "preview_feature_description_month_in_review": "预览月度回顾，现在可在 V3 中直接运行。",
   "preview_feature_description_rewatch": "开始重温您已看过的节目，直接从会话抽屉中签到或标记已观看剧集。",
   "preview_feature_description_scoped_favorites": "通过在个人资料上显示当年的收藏以及按年份分组完整的收藏列表来限制收藏列表的时间范围。",
   "preview_feature_description_smart_recommendations": "使用我们的下一代引擎，发现超越显而易见的推荐电影和节目。",
   "preview_feature_description_smart_related": "使用我们的下一代引擎，发现超越显而易见的电影和节目。",
-  "preview_feature_description_year_in_review": "预览年度回顾，现在可在 V3 中直接运行。",
   "preview_feature_title_edit_mode": "编辑模式",
   "preview_feature_title_leaderboard": "万神殿",
-  "preview_feature_title_month_in_review": "月度回顾",
   "preview_feature_title_rewatch": "重温",
   "preview_feature_title_scoped_favorites": "限定范围的收藏",
   "preview_feature_title_smart_recommendations": "智能推荐",
   "preview_feature_title_smart_related": "智能相关",
-  "preview_feature_title_year_in_review": "年度回顾",
   "reset_status_cleared": "待看列表已清除。",
   "section_title_season_progress": "第 {number} 季 进度",
   "section_title_seasons_overview": "概览",
