+++ projects/client/i18n/meta/en.json
@@ -7951,22 +7951,6 @@
         }
       }
     },
-    "preview_feature_title_year_in_review": {
-      "default": "Year In Review",
-      "description": "Title for the Year In Review preview feature."
-    },
-    "preview_feature_description_year_in_review": {
-      "default": "Preview Year in Review, now running directly in V3.",
-      "description": "Description for the Year In Review preview feature."
-    },
-    "preview_feature_title_month_in_review": {
-      "default": "Month In Review",
-      "description": "Title for the Month In Review preview feature."
-    },
-    "preview_feature_description_month_in_review": {
-      "default": "Preview Month in Review, now running directly in V3.",
-      "description": "Description for the Month In Review preview feature."
-    },
     "preview_feature_title_edit_mode": {
       "default": "Edit Mode",
       "description": "Title for the Edit Mode preview feature."
