+++ projects/client/src/lib/features/feature-flag/models/FeatureFlag.ts
@@ -1,6 +1,4 @@
 export enum FeatureFlag {
-  YearInReview = 'year-in-review',
-  MonthInReview = 'month-in-review',
   EditMode = 'edit-mode',
   ScopedFavorites = 'scoped-favorites',
   SmartRelated = 'smart-related',
